+++ src/expect/src/__tests__/__snapshots__/toThrowMatchers.roblox.spec.snap.lua
@@ -6,8 +6,8 @@ exports[ [=[Lua toThrowMatcher tests cleans stack trace and prints correct files
 
 Thrown value: <r>"attempt to perform arithmetic (add) on nil and number"</>
 
-      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:168 function func2
-      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:174
+      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:169 function func2
+      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:175
 ]=]
 
 exports[ [=[Lua toThrowMatcher tests prints the stack trace for Lua AssertionError error 1]=] ] = [=[
@@ -18,9 +18,9 @@ Error name:    <r>"AssertionError"</>
 Error message: <r>""</>
 
       AssertionError [ERR_ASSERTION]
-      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:78 function error3
-      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:82 function test3
-      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:112
+      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:79 function error3
+      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:83 function test3
+      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:113
 ]=]
 
 exports[ [=[Lua toThrowMatcher tests prints the stack trace for Lua Error error 1]=] ] = [=[
@@ -31,9 +31,9 @@ Error name:    <r>"Error"</>
 Error message: <r>""</>
 
       Error
-      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:62 function error1
-      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:70 function test1
-      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:88
+      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:63 function error1
+      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:71 function test1
+      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:89
 ]=]
 
 exports[ [=[Lua toThrowMatcher tests prints the stack trace for Lua string error 1]=] ] = [=[
@@ -42,9 +42,9 @@ exports[ [=[Lua toThrowMatcher tests prints the stack trace for Lua string error
 
 Thrown value: <r>""</>
 
-      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:66 function error2
-      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:74 function test2
-      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:96
+      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:67 function error2
+      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:75 function test2
+      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:97
 ]=]
 
 exports[ [=[Lua toThrowMatcher tests prints the stack trace for Lua string error 2 1]=] ] = [=[
@@ -54,9 +54,9 @@ exports[ [=[Lua toThrowMatcher tests prints the stack trace for Lua string error
 Expected substring: <g>"wrong information"</>
 Received value:     <r>""</>
 
-      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:66 function error2
-      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:74 function test2
-      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:104
+      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:67 function error2
+      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:75 function test2
+      _Workspace.Expect.Expect.__tests__.toThrowMatchers.roblox.spec:105
 ]=]
 
 exports[ [=[Lua toThrowMatcher tests toThrow should fail if expected is a string and thrown message is a table 1]=] ] =