+++ src/expect/src/__tests__/matchers.roblox.spec.lua
@@ -1,3 +1,4 @@
+--!optimize 1
 --[[
 	* Copyright (c) Roblox Corporation. All rights reserved.
 	* Licensed under the MIT License (the "License");