+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * :hammer_and_wrench: Added an `args` module that parses command line arguments passed in through `roblox-cli` and moving away from relying on globals to pass arguments ([#454](https://github.com/Roblox/jest-roblox-internal/pull/454))
 * :bug: Fix the type for `each` ([#457](https://github.com/Roblox/jest-roblox-internal/pull/457))
+* :hammer_and_wrench: Run CI in OCALE ([#458](https://github.com/Roblox/jest-roblox-internal/pull/458))
 
 
 ## 3.13.3 (2025-06-04)